+++ include/hs_routes_fs.h
@@ -24,9 +24,16 @@ struct HSRoute *hs_routes_fs_directory_route_new(char * /* base directory */);
 /**
  * Returns the directory content.
  * It will ignore non file paths or paths that do not exist.
- * The optional additional_head_content will not be released with the route.
+ * For media entities, such as images and videos, the relevant html elements will be created.
  */
-struct HSRoute *hs_routes_fs_directory_route_new_with_options(char * /* base directory */, char * /* additional_head_content */);
+struct HSRoute *hs_routes_fs_directory_route_new_with_media_support(char * /* base directory */);
+
+/**
+ * Returns the directory content.
+ * It will ignore non file paths or paths that do not exist.
+ * The optional additional_head_content and context will not be released with the route.
+ */
+struct HSRoute *hs_routes_fs_directory_route_new_with_options(char * /* base directory */, char * /* additional_head_content */, bool (*filter)(char * /* name */, bool /* is directory */), char * (*render_directory_entry)(char * /* name */, char * /* href */, void * /* context */), char * (*render_file_entry)(char * /* name */, char * /* href */, void * /* context */), void *context);
 
 #endif