+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## CHANGELOG
 
+### v0.4.0
+
+* Enhancement: New options for the hs_routes_fs_file_route_new_with_options function
+* Enhancement: New hs_routes_fs_directory_route_new_with_media_support function
+
 ### v0.3.0 (2021-09-09)
 
 * Routes support all HTTP methods.